packages/teambition-client/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "teambition-client"
version = "0.1.0"
description = "Teambition Open API client (config-driven, framework-agnostic)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31",
  "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["src"]
